+++ src/app/logs/logs.view.menu.component.ts
@@ -1,8 +1,10 @@
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { UtilSvc } from '../utilities/utilSvc';
 import { UserInfo, CurrentMatch } from '../app.globals';
-import { DataSvc } from '../model/dataSvc';
-import { Match } from '../model/match'
+import { PUBLIC_USER_ID } from '../constants';
+import { DataSvc, PLAYER_TABLE_NAME, TOURNAMENT_TABLE_NAME } from '../model/dataSvc';
+import { Match } from '../model/match';
+import { PlayerData, DEFAULT_GENDER_TYPE } from '../model/player';
 
 @Component({
   selector: '<app-logs-view-menu>',
@@ -11,10 +13,13 @@ import { Match } from '../model/match'
 export class LogsViewMenuComponent implements OnInit, OnDestroy {
 
   @Input() matchMenuOpen    : boolean;    // indicates this panel should open
+  @Input() viewPublics      : boolean = false;  // indicates user is viewing public matches
+  @Input() managePublics    : boolean = false;  // indicates manage public matches feature being used
   @Input() playerName       : Function;   // function to return player name from player id
   @Input() constructMatchesMessage  : Function;  // function to format matchesMessage
 
   selectAll                 : boolean = false; // toggle for selecting/unselecting all matches
+  publicFilter              : string = "Either" // select for which matches are shown in manage public matches
   multiMode                 : string = "Combine";     // mode for handling multiple selections (Combine or Trends)
   menuMatchList             : Match[] = [];    // working list of matches 
   matchSelectFlags          : boolean [] = []; // selection status for each match in matchList
@@ -40,9 +45,9 @@ export class LogsViewMenuComponent implements OnInit, OnDestroy {
   // delete the selected match
   public deleteMatchMenuItem = (index : number) => {
     var m : Match = this.menuMatchList[index];
-    var msg = 'Delete ' + this.playerName(m.playerId) + ' .vs. ' + this.playerName(m.opponentId)+
-            ' on '+ m.date + ' ?';
-    this.utilSvc.getConfirmation('Delete Match', msg, 'Delete')
+
+    this.utilSvc.confirmMatchAction('Delete Match', this.playerName(m.playerId),
+                                    this.playerName(m.opponentId), m.date, 'Delete')
     .then((deleteIt) => {
       this.dataSvc.deleteMatch(m.sortDate)
       .then((success) => {
@@ -80,10 +85,8 @@ export class LogsViewMenuComponent implements OnInit, OnDestroy {
   }
 
   public setSelectedMatch = (index : number) => {
-    if(this.selectedMatchCount() == 0){
-      this.currentMatch.selectedMatches = [this.menuMatchList[index]];
-      this.viewSelectedMatches();
-    }
+    this.currentMatch.selectedMatches = [this.menuMatchList[index]];
+    this.viewSelectedMatches();
   }
 
   // save the objects that have been selected from the match list 
@@ -137,6 +140,211 @@ export class LogsViewMenuComponent implements OnInit, OnDestroy {
     return this.matchSelectFlags[index];
   }
 
+  // make the selected match public
+  public makeMatchPublic = (index : number) => {
+    var m : Match = this.menuMatchList[index];
+    var mpDate : string = m.sortDate;
+
+    // this.utilSvc.confirmMatchAction('Make Match Public', this.playerName(m.playerId),
+    //                                 this.playerName(m.opponentId), m.date, 'OK')
+    // .then((doIt) => {
+    this.utilSvc.openPublicMatchSettings('Add Public Match', undefined, 
+            this.playerName(m.playerId), this.playerName(m.opponentId), m.date, 'Create', 'Make Public')
+    .then((result) => {
+      if(result.create === true){  //are we creating the public copy?
+        if(mpDate.length === 12){ mpDate = this.utilSvc.extendSortDate(mpDate); }
+        this.addPublicMatch(m, mpDate)
+        .then((publicMatchAdded) => {
+          if(result.list !== undefined){
+            this.setEmailRestrictions(mpDate, result.list)
+            .then((success) => {})
+            .catch((failure) => {})
+          }
+          m.madePublic = mpDate;     
+          this.dataSvc.saveMatch(m) // save private version
+          .then((privateMatchUpdated) => {
+            this.utilSvc.displayThisUserMessage("matchMadePublic");   
+          })
+          .catch((failToUpdatePrivateMatch) => {
+            this.utilSvc.displayThisUserMessage("errorUpdatingPrivateMatch");
+            m.madePublic = undefined;  
+          })   
+        })
+        .catch((failToAddPublicMatch) => {
+          this.utilSvc.displayThisUserMessage("errorMakingMatchPublic");
+        })
+      }
+    })
+    .catch((userHitCancel) => {
+    })
+  }
+
+  // store the given match to the database using the PUBLIC_USER_ID
+  private addPublicMatch = (mOrig : Match, sDate : string) => {
+    var m      : Match = Match.build(mOrig.getMatchLog());  //copy the match object
+    var pFirst : string = this.playerName(m.playerId, 'F'),
+        pLast  : string = this.playerName(m.playerId, 'L'),
+        oFirst : string = this.playerName(m.opponentId, 'F'),
+        oLast  : string = this.playerName(m.opponentId, 'L');
+
+    return new Promise((resolve, reject) => {
+      this.dataSvc.getList(PLAYER_TABLE_NAME, PUBLIC_USER_ID)
+      .then((pList) => {
+        // get PUBLIC MATCH ids for the players and replace relevent ids in the match data
+        m.playerId = this.getPublicPlayerId(<PlayerData[]>pList, pFirst, pLast, m.playerHanded);
+        m.opponentId = this.getPublicPlayerId(<PlayerData[]>pList, oFirst, oLast, m.opponentHanded);
+        m.winnerId = mOrig.winnerId == mOrig.playerId ? m.playerId : m.opponentId;
+        m.submittedBy = mOrig.userId;
+        m.userId = PUBLIC_USER_ID;
+        m.sortDate = sDate;
+        this.dataSvc.saveMatch(m)   //save public version
+        .then((success) => {
+          this.dataSvc.saveList(pList, PLAYER_TABLE_NAME, PUBLIC_USER_ID) // save updated PUBLIC players list
+          .then((playerListUpdated) => {
+            this.dataSvc.getList(TOURNAMENT_TABLE_NAME, PUBLIC_USER_ID) // read PUBLIC tournaments list
+            .then((tList) => {
+              this.addTournamentName(<string[]>tList, m.tournament);
+              this.dataSvc.saveList(tList, TOURNAMENT_TABLE_NAME, PUBLIC_USER_ID) // save updated PUBLIC tourn list
+              .then((tournamentListUpdated) => {
+                resolve(tournamentListUpdated);            
+              })
+              .catch((failToSaveTournamentTable) => {
+                reject(failToSaveTournamentTable);
+              })
+            })
+            .catch((failToReadTournamentTable) => {
+              reject(failToReadTournamentTable);
+            })
+          })
+          .catch((failToSavePlayerList) => {
+            reject(failToSavePlayerList);
+          })
+        })
+        .catch((failToSavePublicMatch) => {
+          reject(failToSavePublicMatch);
+        })
+      })
+      .catch((failToReadPlayerList) => {
+        reject(failToReadPlayerList);
+      })
+    })
+  }
+
+  // check player list for given name, return id if found otherwise add name to list and use next id value
+  private getPublicPlayerId = (list : PlayerData[], fName : string, lName : string, handed : string ) => {
+    var maxId : number = 0;
+    var i     : number;
+    var pData : PlayerData = <PlayerData>{};
+
+    for(i=0; i<list.length; i++){   //see if name already in players list
+      if((fName === list[i].firstName) && (lName === list[i].lastName)){
+        return list[i].id;
+      }
+      if(list[i].id > maxId){ maxId = list[i].id };
+    }
+    // name not found, add a new entry to PUBLIC MATCHES players list
+    pData.id = ++maxId;   //create a new id number
+    pData.createdOn = this.utilSvc.formatDate();
+    pData.firstName = fName;
+    pData.lastName = lName;
+    pData.handed = handed;
+    pData.gender = DEFAULT_GENDER_TYPE;
+    list.push(pData);     
+
+    return maxId;
+  }
+
+  // check player list for given name, return id if found otherwise add name to list and use next id value
+  private addTournamentName = (list : string[], tName : string ) => {
+    if(!list.includes(tName)){
+      list.push(tName);
+    }
+  }
+
+  // remove the public copy of the selected match or set authorized users
+  public publicMatchSettings = (index : number) => {
+    var m : Match = this.menuMatchList[index];
+    var sDate = m.madePublic;       //get sortDate for public copy of match
+
+    this.dataSvc.readMatch(PUBLIC_USER_ID, sDate)
+    .then((publicMatch) => {
+      this.utilSvc.openPublicMatchSettings('Public Match Settings', publicMatch.restrictedTo, 
+                      this.playerName(m.playerId), this.playerName(m.opponentId), m.date, 'Edit')
+      .then((result) => {
+        if(result.delete === true){  //are we deleting the public copy?
+          this.dataSvc.deleteMatch(sDate, PUBLIC_USER_ID)
+          .then((success) => {
+            m.madePublic = undefined;
+            this.dataSvc.saveMatch(m)
+            .then((success) => {
+              this.utilSvc.displayThisUserMessage("matchMadePrivate");
+            })
+            .catch((error) => {
+              this.utilSvc.displayThisUserMessage("errorUpdatingPrivateMatch");
+            })
+          })
+          .catch((failure) => {
+            this.utilSvc.displayThisUserMessage("errorDeletingPublicCopy");
+          })
+        } else { // if not deleting, update the authorized users list
+          this.setEmailRestrictions(sDate, result.list)
+          .then((success) => {})
+          .catch((failure) => {})
+        }
+      })
+      .catch((userHitCancel) => {
+      })
+    })
+    .catch((failToReadPublicMatch) => {
+      this.utilSvc.displayThisUserMessage("errorReadingPublicCopy");
+    })
+  }
+
+  // set the email restriction list for the given match
+  setEmailRestrictions = (sDate: string, list: string[]) => {
+
+    return new Promise((resolve, reject) => {
+      this.dataSvc.readMatch(PUBLIC_USER_ID, sDate)
+      .then((publicMatch) => {
+        publicMatch.restrictedTo = list;
+        this.dataSvc.saveMatch(publicMatch)
+        .then((success) => {
+          this.utilSvc.displayThisUserMessage("matchRestrictionsUpdated");
+          resolve(success);
+        })
+        .catch((failToSavePublicMatch) => {
+          this.utilSvc.displayThisUserMessage("errorUpdatingMatchRestrictions");
+          reject(failToSavePublicMatch)
+        })
+      })
+      .catch((failToReadPublicMatch) => {
+        this.utilSvc.displayThisUserMessage("errorReadingPublicCopy");
+        reject(failToReadPublicMatch)
+      })
+    })
+  }
+
+  // test to see if the given match is already public
+  public matchIsPublic = (index : number) => {
+    return this.menuMatchList[index].madePublic;
+  }
+
+  // return whether the match at the given index matches the publicFilter
+  public applyPublicFilter = (index : number) => {
+    var isPublic = this.matchIsPublic(index);
+    return ((isPublic && "Either Public".includes( this.publicFilter)) ||
+            (!isPublic && "Either Private".includes( this.publicFilter)))
+  }
+
+  // update the label on the menu tab
+  public updateMatchCountLabel = () => {
+    var i, count = 0;
+    for(i=this.menuMatchList.length-1; i>=0; i--){
+      count += this.applyPublicFilter(i) ? 1 : 0;
+    } 
+    this.constructMatchesMessage(count);
+  }
+
   // move to the next tab in the tab set
   public nextTab = () => {
     this.utilSvc.emitEvent("nextTab");
@@ -145,6 +353,11 @@ export class LogsViewMenuComponent implements OnInit, OnDestroy {
   // move to the next tab in the tab set
   public prevTab = () => {
     this.utilSvc.emitEvent("prevTab");
+  }
+
+  // close the View Logs display
+  public closeView = () => {
+    this.utilSvc.emitEvent("closeView");
   }
 
 }
